Moved all skipChecker to checker package
